Implement Elliptic curve order, Point addition, Point doubling and double and add algorithm. Cleaned up files in solution

+++ NumberTheory/NT.cpp
@@ -51,7 +51,7 @@ char jacobi(int a, int n)
 int sqrtmodp(int a, int p)
 {
 	a %= p;
-	if (p % 8 == 3 || p % 8 == 7) 
+	if (p % 8 == 3 || p % 8 == 7)
 		return expmod(a, (p + 1) >> 2, p);
 	if (p % 8 == 5) {
 		int x = expmod(a, (p + 3) >> 3, p);
@@ -65,7 +65,7 @@ int sqrtmodp(int a, int p)
 	while (jacobi(d, p) != 1)
 		d = rand() % (p - 2) + 2;
 	int k = 0;
-	int q = p-1;
+	int q = p - 1;
 	while (!(q % 2)) {
 		q >>= 1;
 		k++;
@@ -74,8 +74,8 @@ int sqrtmodp(int a, int p)
 	int D = expmod(d, q, p);
 	int m = 0;
 	for (int i = 0; i < k; i++) {
-		if (expmod(A * expmod(D, m, p), pow(2, k - 1 - i), p) == p - 1)
-			m += pow(2, i);
+		if (expmod(A * expmod(D, m, p), (int)pow(2, k - 1 - i), p) == p - 1)
+			m += (int)pow(2, i);
 	}
 	return expmod(a, (q + 1) >> 1, p) * expmod(D, m >> 1, p) % p;
 }
